+++ packages/expo-app-metrics/CHANGELOG.md
@@ -16,6 +16,8 @@
 
 ### 🐛 Bug fixes
 
+- [iOS] Report `bundleLoadTime` as unavailable on React Native 0.87, which removed the `StartupLogger` accessor it was measured from. ([#47729](https://github.com/expo/expo/pull/47729) by [@gabrieldonadel](https://github.com/gabrieldonadel))
+
 - Fix integer metric and log attributes equal to `0` or `1` serializing as booleans on iOS. ([#47108](https://github.com/expo/expo/pull/47108) by [@tsapeta](https://github.com/tsapeta))
 - fix race condition between db inserts ([#46702](https://github.com/expo/expo/pull/46702) by [@Ubax](https://github.com/Ubax))
 - [tvOS] Fix path for DB creation. ([#46715](https://github.com/expo/expo/pull/46715) by [@douglowder](https://github.com/douglowder))